refactor: moved WarningMessage component to root of components folder
